requirements.txt
requests>=2.32.0
websocket-client>=1.8.0
pyinstaller>=6.0.0
